refactor(button): preserve button width with icon and improve exports
